LayoutTests/platform/gtk/accessibility/name-from-label-expected.txt

Full Name: 
This tests that AtkObjects derive their names from associated labels.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL child.role should be entry. Was AXRole: entry.
FAIL child.title should be Full Name:. Was AXTitle: Full Name:.
PASS successfullyParsed is true

TEST COMPLETE
